Fixing a couple of return types.
git-svn-id: https://svn.resiprocate.org/rep/resiprocate/main@7496 ddefafc4-47db-0310-ae44-fa13212b10f2

+++ repro/Target.hxx
@@ -55,7 +55,7 @@ class Target
       virtual Target* clone() const;
       
       //In case you need const accessors to keep things happy.
-      virtual float getPriority() const;
+      virtual int getPriority() const;
       virtual bool shouldAutoProcess() const;
       
       static bool targetPtrCompare(const Target* lhs, const Target* rhs)
